refactoring(Feed) refactor post component
