+++ src/components/Dashboard/CommunitySection/CommunitySection.component.tsx
@@ -1,7 +1,7 @@
 import React, { memo, useEffect } from 'react';
 import { Outlet, useLocation } from 'react-router-dom';
 import { useQueryClient } from '@tanstack/react-query';
-import { useCommunity } from './hooks/useCommunity';
+import { useCommunity } from './Feed/hooks/useCommunity.hook';
 import { prefetchRanking } from './Ranking/hooks/useRanking';
 import { CommunityNav } from './Feed/components/Comments/CommunityNav.component';
 import { CommunityView } from './types/community.types';
